Separated tag offset input into separate component
