+++ app/javascript/custom-tagged-plate/components/TagSubstitutionDetails.vue
@@ -91,4 +91,5 @@ export default {
     }
   }
 }
+
 </script>
